include: Move private csp_macro.h under src/
csp_macro.h is not a public's API.  Move it under src/.

Because we don't want csp_macro.h to be public, we are not using it
from any of our public headers.
