+++ CMakeLists.txt
@@ -57,8 +57,12 @@ if(CMAKE_SYSTEM_NAME STREQUAL "Linux" AND
 endif()
 
 file(REAL_PATH include csp_inc)
+file(REAL_PATH src csp_inc_src)
 list(APPEND csp_inc ${CMAKE_CURRENT_BINARY_DIR}/include)
-target_include_directories(csp PUBLIC ${csp_inc})
+target_include_directories(csp
+  PUBLIC ${csp_inc}
+  PRIVATE ${csp_inc_src}
+)
 
 if(CSP_POSIX)
   set(CSP_C_ARGS -Wshadow -Wcast-align -Wwrite-strings -Wno-unused-parameter)
