+++ src/arch/posix/csp_clock.c
@@ -1,5 +1,5 @@
 #include <csp/csp_hooks.h>
-#include <csp/csp_macro.h>
+#include "csp_macro.h"
 
 #include <time.h>
