+++ src/csp_debug.c
@@ -1,5 +1,5 @@
 #include <inttypes.h>
-#include <csp/csp_macro.h>
+#include "csp_macro.h"
 #include "csp/autoconfig.h"
 
 uint8_t csp_dbg_buffer_out;
