+++ src/csp_qfifo.c
@@ -3,7 +3,7 @@
 #include <csp/arch/csp_queue.h>
 #include <csp/csp_debug.h>
 #include <csp/csp_buffer.h>
-#include <csp/csp_macro.h>
+#include "csp_macro.h"
 #include "csp/autoconfig.h"
 
 static csp_static_queue_t qfifo_queue __noinit;
