+++ src/csp_sfp.c
@@ -4,7 +4,7 @@
 
 #include <csp/csp_buffer.h>
 #include <csp/csp_debug.h>
-#include <csp/csp_macro.h>
+#include "csp_macro.h"
 #include <endian.h>
 
 #include "csp_conn.h"
